[JENKINS-57173] Strip out invalid characters from XML log output
Darcs stores parts of its logs as is in binary and pump this into
the XML log output unsanitized. This leads to abortion of XML parsing.

This is a workaround and it should be fixed in Darcs itself: There should
be output sanitizing, unless this bug would lead to DoS of consuming
software.

This patch is submitted by Ilya Perminov <iperminov@dwavesys.com>